Comply with rubocop's instructions
