+++ spec/lib/git/duet/cli_spec.rb
@@ -46,46 +46,46 @@ describe Git::Duet::Cli do
   end
 
   it 'defaults to loud and local when running `solo`' do
-    allow(Git::Duet::SoloCommand).to receive(:new).with('jd', be_falsey, be_falsey).and_return(
-      double('solo').as_null_object
-    )
+    allow(Git::Duet::SoloCommand)
+      .to receive(:new).with('jd', be_falsey, be_falsey)
+      .and_return(double('solo').as_null_object)
     cli.run('git-solo', %w(jd))
   end
 
   it 'parses options for quietness and globality when running `solo`' do
-    allow(Git::Duet::SoloCommand).to receive(:new).with('jd', be_truthy, be_truthy).and_return(
-      double('solo').as_null_object
-    )
+    allow(Git::Duet::SoloCommand)
+      .to receive(:new).with('jd', be_truthy, be_truthy)
+      .and_return(double('solo').as_null_object)
     cli.run('git-solo', %w(jd -q -g))
   end
 
   it 'defaults to loud and local when running `duet`' do
-    allow(Git::Duet::DuetCommand).to receive(:new).with('jd', 'fb', be_falsey, be_falsey).and_return(
-      double('duet').as_null_object
-    )
+    allow(Git::Duet::DuetCommand)
+      .to receive(:new).with('jd', 'fb', be_falsey, be_falsey)
+      .and_return(double('duet').as_null_object)
     cli.run('git-duet', %w(jd fb))
   end
 
   it 'parses options for quietness and globality when running `duet`' do
-    allow(Git::Duet::DuetCommand).to receive(:new).with('jd', 'fb', be_truthy, be_truthy).and_return(
-      double('duet').as_null_object
-    )
+    allow(Git::Duet::DuetCommand)
+      .to receive(:new).with('jd', 'fb', be_truthy, be_truthy)
+      .and_return(double('duet').as_null_object)
     cli.run('git-duet', %w(jd fb -q -g))
   end
 
   it 'respects the environment about global issues when running `solo`' do
     stub_const('ENV', 'GIT_DUET_GLOBAL' => 'true')
-    allow(Git::Duet::SoloCommand).to receive(:new).with('cp', anything, be_truthy).and_return(
-      double('solo').as_null_object
-    )
+    allow(Git::Duet::SoloCommand)
+      .to receive(:new).with('cp', anything, be_truthy)
+      .and_return(double('solo').as_null_object)
     cli.run('git-solo', %w(cp))
   end
 
   it 'respects the environment about global issues when running `duet`' do
     stub_const('ENV', 'GIT_DUET_GLOBAL' => 'true')
-    allow(Git::Duet::DuetCommand).to receive(:new).with('cp', 'g', anything, be_truthy).and_return(
-      double('duet').as_null_object
-    )
+    allow(Git::Duet::DuetCommand)
+      .to receive(:new).with('cp', 'g', anything, be_truthy)
+      .and_return(double('duet').as_null_object)
     cli.run('git-duet', %w(cp g))
   end
 end
